Start moving flyway into backend
