+++ back-end/src/main/java/dev/davidsilva/music/database/FlywayConfiguration.java
@@ -1,0 +1,26 @@
+package dev.davidsilva.music.database;
+
+import org.flywaydb.core.Flyway;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class FlywayConfiguration {
+    private final DataSource dataSource;
+
+    public FlywayConfiguration(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    @Bean
+    public Flyway flyway() {
+        return Flyway
+                .configure()
+                .dataSource(dataSource)
+                .locations("db/migration")
+                .cleanDisabled(false)
+                .load();
+    }
+}
